+++ finance/urls.py
@@ -7,6 +7,10 @@ urlpatterns = [
     path("estimate/<e_id>", views.viewEstimate.as_view(), name="estimate"),
     path("invoice/<e_id>", views.viewInvoice.as_view(), name="invoice"),
     path("timesheet/<pp_id>/<emp_id>", views.viewTimesheet.as_view(), name="timesheet"),
+    path("noprint/timesheet/<pp_id>/<emp_id>", views.ViewTimesheetNoPrint.as_view(), name="timesheetnoprint"),
+    path(
+        "sign/timesheet/<timesheet_id>", views.SignTimesheet.as_view(), name="sign_timesheet"
+    ),
     path(
         "rollover/<pp_id>/<emp_id>", views.RollOverAllShifts.as_view(), name="rollover"
     ),